Improved efficiency.
a(n) now tests to see if A000188(n)=1, and if it does, applies recurrence relation a(n) = a(n-1) + 1 and calculates a(n-1) via brute force instead. This reduces the testing amount from 3^n to 3^(n-1). More efficiency can be still be gained, however I have yet to rewrite the list function to allow for this.